Move the constants.js file to /helpers
