+++ ui/job-view/details/tabs/failureSummary/SuggestionsListItem.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import { thBugSuggestionLimit } from '../../../../js/constants';
+import { thBugSuggestionLimit } from '../../../../helpers/constants';
 import BugListItem from './BugListItem';
 
 export default class SuggestionsListItem extends React.Component {
