+++ ui/js/controllers/perf/compare.js
@@ -8,7 +8,7 @@ import {
   phCompareDefaultNewRepo,
   phTimeRanges,
   phCompareBaseLineDefaultTimeRange,
-} from '../../constants';
+} from '../../../helpers/constants';
 import PushModel from '../../../models/push';
 import RepositoryModel from '../../../models/repository';
