+++ ui/models/job.js
@@ -1,4 +1,4 @@
-import { thPlatformMap } from '../js/constants';
+import { thPlatformMap } from '../helpers/constants';
 import { createQueryParams, getProjectUrl } from '../helpers/url';
 import { formatTaskclusterError } from '../helpers/errorMessage';
 import TaskclusterModel from './taskcluster';
